+++ bin/quads-server
@@ -18,33 +18,14 @@
 import argparse
 import cherrypy
 import os
-import yaml
 import sys
-import logging
-import threading
-import re
-import cgi
-import json
-import threading
+from quads.api_v2 import QuadsServerApiV2
+from quads.helpers import quads_load_config
 
 
 # TODO: is there a better way to handle these constants??
 # TODO: revisit loading configfile and getting defaults
 #       vs getting overrides from the user
-# load quads config
-
-def quads_load_config(quads_config):
-    try:
-        with open(quads_config, 'r') as config_file:
-            try:
-                quads_config_yaml = yaml.safe_load(config_file)
-            except Exception as ex:
-                print("quads: Invalid YAML config: " + quads_config)
-                exit(1)
-    except Exception as ex:
-        print(ex)
-        exit(1)
-    return(quads_config_yaml)
 
 conf_file = os.path.join(os.path.dirname(__file__), "../conf/quads.yml")
 conf = quads_load_config(conf_file)
@@ -54,11 +35,6 @@ defaultstatedir = os.path.join(conf["data_dir"], "state")
 sys.path.append(conf['install_dir'])
 sys.path.append(os.path.dirname(__file__) + "/../")
 
-#from quads.webui import QuadsWebUI
-#from quads.api_v1 import QuadsServerApiV1
-from quads.api_v2 import QuadsServerApiV2
-#from quads.util import quads_load_config
-
 
 class QuadsServer(object):
     @cherrypy.expose
@@ -66,12 +42,12 @@ class QuadsServer(object):
         return 'index.html'
 
 
-#https://docs.cherrypy.org/en/latest/advanced.html#restful-style-dispatching
+# https://docs.cherrypy.org/en/latest/advanced.html#restful-style-dispatching
 @cherrypy.expose
 class QuadsServerApi(object):
     def __init__(self):
         # v1 doesn't work, this is just an example
-        #self.v1 = QuadsServerApiV1()
+        # self.v1 = QuadsServerApiV1()
         self.v2 = QuadsServerApiV2()
 
 
@@ -108,7 +84,7 @@ def main():
 
     # Cherrypy configurations
     static_path = os.path.abspath(
-                      os.path.join(os.path.dirname(__file__), 'static'))
+        os.path.join(os.path.dirname(__file__), 'static'))
     cherrypy.config.update({'server.socket_host': args.host,
                             'server.socket_port': args.port,
                             'tools.staticdir.root': static_path})
@@ -139,11 +115,12 @@ def main():
     }
     cherrypy.tree.mount(webapp, '/', conf)
 
-    #cherrypy.quickstart(StringGeneratorWebService(), '/', conf)
+    # cherrypy.quickstart(StringGeneratorWebService(), '/', conf)
     # Start up CherryPy
     cherrypy.engine.signals.subscribe()
     cherrypy.engine.start()
     cherrypy.engine.block()
+
 
 if __name__ == "__main__":
     main()
