+++ quads/helpers.py
@@ -1,0 +1,34 @@
+import yaml
+
+
+# Load QUADS yaml config
+def quads_load_config(quads_config):
+    try:
+        with open(quads_config, 'r') as config_file:
+            try:
+                quads_config_yaml = yaml.safe_load(config_file)
+            except Exception as ex:
+                print("quads: Invalid YAML config: " + quads_config)
+                exit(1)
+    except Exception as ex:
+        print(ex)
+        exit(1)
+    return quads_config_yaml
+
+
+def param_check(data, params, defaults={}):
+    result = []
+    # set defaults
+    for k, v in defaults.items():
+        data.setdefault(k, v)
+
+    if data:
+        # check for missing params
+        for p in params:
+            if p not in data:
+                result.append("Missing required parameter: %s" % p)
+            elif not (data[p] or data[p] is None):
+                result.append("Could not parse %s parameter" % p)
+            elif data[p] == 'None':
+                data[p] = None
+    return result, data
